+++ mods/BrewLAN/modules/BrewGantry3/hook/mods/BrewLAN/lua/classes/Gantry.lua
@@ -13,27 +13,12 @@ do
                 totalmass = totalmass + brain:GetArmyStat("Economy_TotalProduced_Mass", 0.0).Value - brain:GetArmyStat("Economy_AccumExcess_Mass", 0.0).Value
             end
         end
-        --LOG("Total enemy mass = " .. totalmass)
         return totalmass
     end
 
-    local OldBrewLANExperimentalFactoryUnit = FactoryUnit
+    local OldBrewLANExperimentalFactoryUnit = ExperimentalFactoryUnit
 
-    do
-        local ok, retclass = pcall(function()
-            --if BrewLANExperimentalFactoryUnit then
-                return BrewLANExperimentalFactoryUnit
-            --end
-        end)
-
-        if ok then
-            OldBrewLANExperimentalFactoryUnit = retclass
-        else
-            WARN("Gantry Hax module loaded before BrewLAN.")
-        end
-    end
-
-    BrewLANExperimentalFactoryUnit = Class(OldBrewLANExperimentalFactoryUnit) {
+    ExperimentalFactoryUnit = Class(OldBrewLANExperimentalFactoryUnit) {
 
         AIStartCheats = function(self)
             ------------------------------------------------------------------------