+++ mods/BrewLAN_Plenae/CityGeneration/units/SEB5101/SEB5101_script.lua
@@ -1,7 +1,16 @@
 --------------------------------------------------------------------------------
 -- UEF Wall: With cordinal scripting
 --------------------------------------------------------------------------------
-local BrewLANCardinalWallUnit = import('/lua/defaultunits.lua').BrewLANCardinalWallUnit
+local BrewLANCardinalWallUnit = import('/lua/defaultunits.lua')
+
+do
+    local ok, retclass = pcall(function() return BrewLANCardinalWallUnit.BrewLANCardinalWallUnit end)
+    if ok then
+        BrewLANCardinalWallUnit = retclass
+    else
+        BrewLANCardinalWallUnit = BrewLANCardinalWallUnit.WallStructureUnit
+    end
+end
 
 SEB5210 = Class(BrewLANCardinalWallUnit) {}